requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0
django-allauth==0.34.0
django-bootstrap4==0.0.5
django-richtextfield==1.2.1
django-smart-selects==1.5.3
idna==2.6
mysqlclient==1.3.12
oauthlib==2.0.6
Pillow==5.0.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
